feature: DataStore

+++ pkg/app/cmd_dl.go
@@ -7,11 +7,18 @@ import (
 
 	"gitar/pkg/client"
 	"gitar/pkg/config"
+	"gitar/pkg/data"
 	"gitar/pkg/utils"
 	"github.com/sirupsen/logrus"
 )
 
-func DownloadArchive(url string) error {
+func DownloadArchive(url string, shouldSendMail bool) error {
+	err := DoDownloadArchive(url, shouldSendMail)
+	logrus.Infof("All done")
+	return err
+}
+
+func DoDownloadArchive(url string, shouldSendMail bool) error {
 	logrus.Infof("Download archive")
 
 	cfg, err := config.LoadConfig()
@@ -33,28 +40,76 @@ func DownloadArchive(url string) error {
 	}
 	logrus.Infof("Archive: %+v", *arc)
 
+	err = os.MkdirAll(cfg.DataDir, os.ModePerm)
+	if err != nil {
+		return err
+	}
+
+	store := data.NewSqlite3DataStore(filepath.Join(cfg.DataDir, "gitar.sqlite"))
+	err = store.Open()
+	if err != nil {
+		return err
+	}
+
+	repoKey := fmt.Sprintf("%s:%s/%s", repoUrl.Platform, repoUrl.Owner, repoUrl.Repo)
+	err = store.SaveRepo(repoKey)
+	if err != nil {
+		return err
+	}
+
+	markDownloaded, err := store.IsCommitDownloaded(arc.Commit)
+	if err != nil {
+		return err
+	}
+
+	if markDownloaded && !shouldSendMail {
+		logrus.Warnf("Commit already downloaded: %s", arc.Commit)
+		return nil
+	}
+
 	arcFile := fmt.Sprintf("%s.tar.gz", arc.Name)
 	destDir := filepath.Join(cfg.RepoDir, repoUrl.Platform, repoUrl.Owner, repoUrl.Repo)
 	destPath := filepath.Join(destDir, arcFile)
-	// TODO 判断是否已存在
 
 	tempFile := fmt.Sprintf("%s-%s.tar.gz", arc.Name, arc.Commit)
 	tempPath := filepath.Join(cfg.TempDir, tempFile)
-	// TODO 删除旧文件
-	err = utils.Aria2Download(arc.TarUrl, cfg.TempDir, tempFile)
-	if err != nil {
-		return err
-	}
-	err = os.MkdirAll(destDir, os.ModePerm)
+	logrus.Infof("Temp file: %s", tempPath)
+
+	destExists, err := utils.FileExists(destPath)
 	if err != nil {
 		return err
 	}
 
-	logrus.Infof("Move file %s => %s", tempPath, destPath)
-	err = os.Rename(tempPath, destPath)
-	if err != nil {
-		return err
+	if destExists {
+		logrus.Warnf("Already downloaded: %s", destPath)
+	} else {
+		err = os.RemoveAll(tempPath)
+		if err != nil {
+			return err
+		}
+
+		err = utils.Aria2Download(arc.TarUrl, cfg.TempDir, tempFile)
+		if err != nil {
+			return err
+		}
+
+		err = os.MkdirAll(destDir, os.ModePerm)
+		if err != nil {
+			return err
+		}
+
+		logrus.Infof("Downloaded: %s", destPath)
+		err = os.Rename(tempPath, destPath)
+		if err != nil {
+			return err
+		}
 	}
+
+	if !markDownloaded {
+		err = store.SetCommitDownloaded(arc.Commit)
+	}
+
+	// TODO send mail
 
 	return err
 }
